Update Ammo_AmmunitionBox_Nuke_ArrowIV.json
This is mostly a description update to remove potential confusion about which weapons can use this box. But I noticed that CASE could negate a 9000 point ammo explosion which seemed absurd. So i reduced it to normal for an Arrow IV box and added fluff explaining that the nuke would not actually go off if the box was destroyed. Since it actually doesn't anyway.