Artillery Support
Reworked the Thumper Urbie as it was in the wrong weight bracket (35 tons instead of 30) with the OP's stats. Created the 200X based off of this, it mounts a sniper artillery piece and 3 tons of ammo. Created the R71 with 4 flamers and 2 MML5's. Made the R60LRM with an LRM-20 and 2 medium lasers. R60LRM and R100X are conversions of custom mechs by Bishop Steiner using the stats he posted online.