Proposed Hydra Nerf
Hydra LRM ammo is the defacto LRM choice as much as Ace Pilot was the default pilot skil choice.

Because of the way it's stats work it basically deals full damage even if it misses. 2 dmg 2 heat in AOE. Even more if they actually hit.

This scales it back so it's not putting out more damage/heat than the base damage of an LRM. So it's scaled like Incendiary, or Swarm.

Stats don't have to be this exactly. But I think something close to this would be more balanced.