Indirect Fire closer to TT
Tabletop is +1 to hit indirect. We have +8 (!!!)  Based on cursory research we set this penalty of 8 before LowVis even existed.  I think it was an attempt to aproximate double-blind hit % with the one value we could change. But now we've got sensor penalties on top of this so it nerfs LRM into the ground for most of the game. Based on personal testing the game just plays better with closer to TT penalty. Half of turns aren't taken up by volleys of LRM that do literally nothing. With it set to 3 max tactics pilots will still have the TT +1 to hit. With a dedicated TTS you can actually get a bonus to hit indirect. But that's fine. It's not like there aren't direct fire gear that give huge bonuses.