Pilot traits description fix - To-Hit tweak
The accuracy change per skill point isn't a change. As far as i can tell our trait description was just wrong and this corrects it to what it's really been for a long time.

Bumping base to-hit to 80% is something we discussed as it's closer to TT. If i recall we were at least gonna try it. Putting it in dev to see how it goes