Event edits
Added the tag used to roll magpies event to the exclusions for the preceding event. Should prevent it from respawning/restarting the chain if the correct options are picked.  Added a requirement tag to the final event so it won't spawn if Magpie is in party at start of game. Added a 50/50 choice to the last event so opening the pod can give 1 wound instead of loot