AP Gauss Battery Refresh
I've adjusted the AP Gauss Batteries to be closer to  8 AP Gauss Rifles as it was the original intention. It seems that the Battery missed a couple of changes to the AP Gauss Rifles over it's lifetime. The damage profile is now in line with AP Gauss and I've bumped the slots up so you are getting only a 25% slot savings versus the 50%. I've also updated the unique Juggernaut that has these as fixed equpiment to make room for the additional slots.