Update description text in Pulse Laser Module
The current functionality of the canon pulse laser module can't be easily replicated due to the way that weapon categories are set up. This description could be slightly misleading, so I've included a plausible non-canon explanation for this discrepancy that can be changed back later if the issue is ever addressed further down the line.