Change LanceSight2 and LanceSight3 to AllLanceSight
Here is a more controversal change. As far as I can tell there is no
difference at all in the bonus LanceSight2 and LanceSight3 and the
only difference between the bonuses LanceSight2 and LanceSight3 and
AllLanceSight is that LanceSight2 and LanceSight3 add a fixed amount
to the base value were AllLanceSight multiplies the base value with
a multiplier. I think this difference can be displayed by the
different formating of the values. But if that is not the case, let
me know what you prefer and I rework the commit.