Fix Weapon_Laser_BeholderStare
The Weapon has no aura therefore the bonus description should be
LanceResolve2 instead of LanceResolve. Also, the values for Painter
and LanceResolve2 are 2 instead of 1.

Now to the more controversial side of my change. The bonus
description indicated, that the weapon was a Bombast Laser but the
weapon has no Modes at all. The weapon description indicated, that
the weapon derivates from a Large Pulse Laser but the weapon has
neither an accuracy or evasion pip ignore bonus nor multiple shots
when fired. As the weapon is as far as I can tell most in line with
a Large Laser I changed everything accordingly.