zee values they do nothing!
We can talk about the others. But these range manipulation values on arty are going to do one of two things: nothing, or weirdness. Depending on circumstance. The Modes are already a mess of Range shenenegins. And malusing range from 3000 to like 2000 is pointless on the Arty, and from like 600 to 400 a big oof on the cannons. Let the Gauss sized stuff shoot at Gauss range!