Deja Vu and minor fixes
There Longtom is a train of six vehicles. We had only the 65 ton main component. Which I thought I fixed ages ago. Here's the stats for the whole shebang represented as well as possible. It's missing a few tons of ammo but still has more than enough for RT engagements.