fixing bad assembly variant fixes
There were some errors in a redbat tool run a couple days ago that squashed some different vehicles together.

The Shepard tank series are not Bulldogs but ya know what...i'm kinda ok with them combining, they're close. I'll allow it.

Padilla and Padilla Heavy are NOT the same. Different eras, motive type and 20 tons between them. Only same name to be confusing.

Fixed addition of typo to PEGASUS_SRM assembly variant.

The meme Navi VTOL is not a Yellowjacket at all.

SM2 Arty is not an SMTankDestroyer. There's a real mess of things named SM#. The major groups are a clan hovertank family and a tracked arty platform.

Vel Harasser is not a Vedette. 25 ton hover not 50 ton tracks. (It's also not a IS Harasser for the record)