LAC10 LAC20 weight fix
They were significantly too heavy relative to the ratios seen for the 2 and 5 light versions to the standard.

Weight of LAC10, LAC20 weapons reduced, ammo loads tweaked to compensate on units equiped with them.

I THINK the uberjager has enough space in the ST for the added ammo. I had to put 2 in the arms...

HBK and Kobold change should be uncontroversial, upgrade from Caseless to Doubles.

The rest are mine.