We have so many rocket launchers
I would not have started this if I'd realized.

Small buff to tandem and aoe normal damage. +1 to heat impair on inferno. all of which were underperforming imo.

Fixed a bunch of off-spec tac values. FIxed a bunch of bolt-on malus. Pretty sure it's -1 not -2. And some things were missing it entirely.