Chappy quirk
Chaparral have an unusal design for Shoot and Scoot with the majority of their armor on the rear. Mechancially that's very hard to use properly in our system since they have to face the enemy after moving to shoot. I've given them 360 firing arc as a workaround. So they can point their butts at the enemy and still shoot.