Smol Guided/Shaped arty buff
oops. wrong missile HE on the Bomb AAA Arrow. Led me to contemplating that 100 pinpoint is less than the total damage of a standard Arrow. Thought about buffing it to 120. But decided to give it 100 dmg and 20 AOE instead of the 20 pinpoint and 100 AOE of standard arrow. Then did the other arty for consistency.

I'm not married to this. Will roll back on good argument.