Update Commander EI/PS events
Update the Commander portion of the EI and PS events to prevent player selecting Commander when the pilot is in recovery after having EI or PS operation.

Added "commander_timeout" temporary tag as DIBILITATED temporary tag does not prevent event triggering for Commander.

event will still fire, but option to do procedure will be unselectable and player is forced to choose the WAIT option