Replace Useless Vee quirk
Easy to Maintain doesn't currently do anything for vehicles.
I replaced it on everything where I'd put it on by choice.
Added another quirk on ones that have it canonically.
If multiple quirks are ever a problem that can be pruned.
Note the Vector, not touched here, canonically has both Easy Maint and Imp Comms.