Ivan's weapons revision
Checked for it doing up to date values for all the fluids at once.
Strictly it should do 40 damage if it's firing Acid and Inferno but that seems way excessive.

Range tweak is a buff but it's also now following the standard step pattern of BT weapons. With 30 meters longer base range than a normal fluid gun.

Gave it limited internal ammo of 20 shots instead of being flamer-like. There isn't a fluid gun internal ammo but chemlaser should be fine?

CACIncomingHeatMult is the new'ish Oil effect minus the heat diminishing returns effect.