Fury II re-researched
Info about the export model Fury II is a bit fuzzy, just mentioned in the fluff of an ancient TRO with out of date stats. But I noted that the originals canonically had an M9 not an M7 Gauss. Making that switch lets us give it a decent ammo count and the 1 ton infantry bay the original Fury had