Update itemCollection_Mechs_Starting_Militia_Techie.csv
This list probably isn't going to be used by Milita start anymore. But several other starts use it so it's maintained with some updates. It mostly will give Technicals the vast majority of the time. If it rolls a Small one it gives 2 copies. The player can get lucky (?) and has a small chance of rolling a few other basic light vehicles.