Big Long Tom Rebalance - It's hard to get one vee and 5 trailers right as just one vee...
This represents an entire six unit train of the self propelled gun, two support trailers, and 3 ammo trailers. The 95 ton total is fudged up to 100 because I just can't fit everything it's supposed to have on paper. The same with fudging the standard armor to ferro. I can't get enough points for it all. Added (empty) turret to vehicle to get one step closer being multiple objects that could be hit. Cargo bay and BA bay removed, it was based on a misunderstanding, tonnage used to up-armor. Getting us near the collective armor protection of the whole conglomerate. but in 5 locations instead of 6 vehicles...