"unit_singleton"
Tag for units with no variants to give them an extra entry in the spawn pool. (added tag to a ton of vehicles and a few mechs)

Also -3 to "unit_legendary" to tamp down a bit on ones boosted by other tags. Replaced "unit_MobileHQ" with more general "unit_apc" on the list.