Reshuffle active lances in Escort
I know it's currently disabled even if it stays that way this might be useful in other ormats of Escort.

Turned off the Hunter Lance that spawns when the convoy reaches the final point. Turned back on the Ambush lance. Set it up so I think the contract will just END when the convoy extracts.

The player will fight a blocking force and a support lance at the start. Then face an ambush. Then it will mission complete if the convoy makes it. Killing all enemies is optional if players want to attempt wacky pure objective mobility play.