Indirect Fire Bonus Stack Limit
All instances deliver a bonus +1 at a time on the same ID. There is a stack limit of 3. Does this work? I have concerns about multiple instances of the same effectID in the same StatusEffect list. This is not done other places as far as I know.

The purpose is to prevent a positive bonus to indirect fire that makes it better than direct fire.