SAR: Radical - Literally no one is paying you.
This is on the company's dime to get your pilot back. Players will hate it. Good. It's real simgame pressure to play it cheaply. And I'll continue tweaks to make that possible in ways that aren't the 1.5 turn LAM/VTOL cheese. Secondary objectives will give actual cash early on, and somewhat offset drop/ammo costs in late game.