An audit of Clan weapon prices
This is consistently double book value and/or the price of the IS equivilent. I think that was actually the plan at some point. Or at any rate they were supposed to be more expensive than IS weapons. Which in many cased currently they were not. We've got consistency now at least.

I do note that this does make some clan items very valuable as cash targets in salvage. But also, running clan items will be a MUCH bigger difficulty spike on Company difficulty.