Vee AI tweaks
Increase in weight to VTOL desire to face enemies.

Quick and dirty increase in desire for vehicles to move as potential bandaid to some issues just in case the reason it seems like they derp more often than mechs is bad values in global AI constants. Not always an AI bug per say.