Update fail chances on the radical pirate kits so the traits match the actual behavior.
Now that KMission has clamped the displayed fail chance to 0% on turns when a failure roll isn't actually made,
we can make failure chances actually accurate, and accurately displayed, for emergency cooling systems.

Most such gear has `FailRoundsStart: 2`, which means the failure can only start to roll on the second activation
that _begins_ with the equipment active. `FlatFailChance` continues to accumulate during this time,
so the initial failchance should be set to a value such that it will accumulate to the desired
initial fail chance, on the first allowed failure round.

As an example, the pirate PDHS Kit changed here has FailRoundsStart: 2.
We want its first failure roll to be 40%, and escalate rapidly (+50% per turn).
Thus, its initial chance is set at -60%. This is the failure chance on Round 0 (the moment of activation).
On Round 1 (first round that begins) with the kit active, the chance accumulates to -10%.
However, no roll is made this round (even if the chance were positive).
On Round 2, the chance accumulates to 40%. The first failure roll will be made this round.