Update event_solarisSeasonStart4.json
Stripped out unneeded exclusion tags and fixed an oversight. Event added   "company_solaris_championship_ticket" when fp dialogue was started but if player declines then the tag stays. The conversation preceding this had the tag as an exclusion so if the player launched and then declined it would block any further attempts at progress