AC5 etc Heat
Standardized on 7 heat per shot because with our buffs they do almost as much damage as a gosh darned TT AC10. (Which i currently have at a nominal 9 heat) Also buffed LAC2 and 5 damage to base for the type. It feels petty to ding them and not the 10 and 20. I'm half  tempted to argue for removing or reducing the AC damage buffs we inherited from HBS but this is reasonable heat for now.