Perf: Improve AI perf - StratOps edition StratOps was constantly checking unit tags for a resupply tag during AI operations, TagSet checks are slow because they first clone the TagSet (in the manner they were being accessed), which is slow and then perform a contains operation, which is also somewhat slow on a Set.
This changes moves these checks to a boolean stat, which is a faster lookup/comparison than a string compare and doesn't need to clone TagSets.
Also removes some redundant list filtering that was being done, removing extra checks for a unit being dead and list iterations.

This should ideally result in a noticable AI speed up