Perf: Improve AI perf - IRBT Edition
IRBTModUtils' Movement Feature was using a TagSet check to determine if a unit is immobile (for things like Overlords), Fetching the tagset is very expensive in a method like this that is frequently called as part of AI calculations.

A Stat version of this tag already exists, by checking the tags at initialization time of the stat we can correctly set its value there and then only check the stat from that point onward, increasing AI think performance.