Cluster Hit weapon crit/tac adjustements
Curving crit chances to account for every shot hits, making crit chance similar to what you'd expect in TT mechanics. This can be tuned later. Missiles gained some max armor thickness. The concept is that it's the 25 damage one "cluster" would be in TT, multiplied by the 5*Damage we use for missiles to calc max.