BA missile AMS bonus removed
I think as unimaginative devs we felt like BA stuff needed a boost. Then players got their hands on it and found the exploitables. Turns out BA missiles are good and don't need all the buffs we gave them. Probably going to leave their +1 HP for old time's sake. But they should be more in line now with the damage nerf and reduction in AMS resist. MRM keep a little because the big boy ones have it too