Flamer back to basics
We're keeping both damage and heat, it's an official optional rule.

But to corespond with the MG changes I want to reduce them to canonical range. Also taking away their falloff to compensate.
I've been using mech mgs and a lot of BA in testing and flamers having boosted range felt unbalanced. I think the alternate option is keeping the range boost and make the falloff rougher. it feels like it barely applies, something might be wrong there.