chem lasers
IS: 3/4 weight and 3/4 heat. Clan 2/4 weight and 4/4 heat.
Nerf from 3/4 weight 1/2 heat for IS, and 2/4 weight 3/4 heat for Clan
I think we slept on how powerful the lower weight we gave them back in the day was.
The TT Ones should probably be cooler than standard lasers if they need ammo. But we made ours both cooler and lighter, which is a lot.