"Clustering" Description update
HBS Clustering, their mechanism to make hits more likely to strike a single location is now referred to as "Grouping" in player facing descriptions. This is to avoid confusion with our pseudo cluster table weapons, lbx cluster rounds etc. Too many Clusters, and Group might actually be a better term for this one.